+++ tolicodes.com-frontend/src/data/data.js
@@ -1,0 +1,14 @@
+import { menuLinks } from './menuLinks';
+import { socialMediaLinks } from './socialMediaLinks';
+import { siteName } from './siteName'; 
+import { description } from './description'; 
+import { youTubeVideo } from './youTubeVideo';
+
+
+export const data = {
+  menuLinks: menuLinks,
+  socialMediaLinks: socialMediaLinks,
+  siteName: siteName,
+  description: description,
+  youTubeVideo: youTubeVideo,
+}
