+++ tolicodes.com-frontend/src/data/menuLinks.js
@@ -1,0 +1,26 @@
+export const menuLinks = [
+  {
+    title: "Resume",
+    url: "/resume",
+  },
+  {
+    title: "CodeMentor",
+    url: "/codeMentor",
+  },
+  {
+    title: "My story",
+    url: "/my-story",
+  },
+  {
+    title: "Services",
+    url: "/services",
+  },
+  {
+    title: "Hobbies",
+    url: "/hobbies",
+  },
+  {
+    title: "Open Source",
+    url: "/open-source",
+  },
+]
