+++ tolicodes.com-frontend/src/data/siteName.js
@@ -1,0 +1,4 @@
+export const siteName = {
+  siteLogoSrc: 'Logo Goes Here',
+  sitTagLine: 'Anatoliy D. Zaslavskiy',
+};
