+++ tolicodes.com-frontend/src/data/socialMediaLinks.js
@@ -1,0 +1,12 @@
+import { SocialNetworkType } from '../components/SocialNetworkMenu'
+
+export const socialMediaLinks = [
+  {
+    type: SocialNetworkType.LinkedIn,
+    url: "https://www.linkedin.com/in/tolicodes/",
+  },
+  {
+    type: SocialNetworkType.GitHub,
+    url: "https://github.com/tolicodes/tolicodes.com",
+  },
+];
