+++ tolicodes.com-frontend/src/data/youTubeVideo.js
@@ -1,0 +1,5 @@
+export const youTubeVideo = {
+  src: "https://www.youtube.com/embed/z1VNwapOb8E",
+    height: '501px',
+      title: 'Watch My Story'
+}
